refactor(ng-add): check if hammer is already present in workspace/main.ts #3256
