+++ projects/igniteui-angular/schematics/utils/dependency-handler.ts
@@ -1,10 +1,8 @@
 import { Tree } from '@angular-devkit/schematics/src/tree/interface';
-import { SchematicContext, Rule, SchematicsException } from '@angular-devkit/schematics';
+import { SchematicContext, Rule } from '@angular-devkit/schematics';
 import { WorkspaceSchema } from '@angular-devkit/core/src/workspace';
 import { getWorkspace } from '@schematics/angular/utility/config';
 import { Options } from '../interfaces/options';
-
-const hammerJsMinAddress = './node_modules/hammerjs/hammer.min.js';
 
 function logIncludingDependency(context: SchematicContext, pkg: string, version: string): void {
     context.logger.info(`Including ${pkg} - Version: ${version}`);
@@ -29,24 +27,7 @@ export function addDependencies(options: Options): Rule {
         const dependencies = 'dependencies';
         const devDependencies = 'devDependencies';
 
-        Object.keys(pkgJson.dependencies).forEach(pkg => {
-            const version = pkgJson.dependencies[pkg];
-            switch (pkg) {
-                case 'hammerjs':
-                    logIncludingDependency(context, pkg, version);
-                    addPackageToPkgJson(tree, pkg, version, dependencies);
-
-                    // import hammerjs in the main.ts file
-                    const mainTsPath = 'src/main.ts';
-                    const contents = 'import \'hammerjs\';\n' + tree.read(mainTsPath).toString();
-                    tree.overwrite(mainTsPath, contents);
-                    break;
-                default:
-                    logIncludingDependency(context, pkg, version);
-                    addPackageToPkgJson(tree, pkg, version, dependencies);
-                    break;
-            }
-        });
+        includeDependencies(pkgJson, context, tree, dependencies);
 
         // Add web-animations-js to dependencies
         Object.keys(pkgJson.peerDependencies).forEach(pkg => {
@@ -64,11 +45,71 @@ export function addDependencies(options: Options): Rule {
 }
 
 /**
- * Add Hammer script to angular.json section
+ * Recursively search for the first property that matches targetProp within the angular.json file.
+ */
+export function getPropertyFromWorkspace(targetProp: string, workspace: any, curKey = ''): any {
+    if (workspace.hasOwnProperty(targetProp)) {
+        return { key: targetProp, value: workspace[targetProp] };
+    }
+
+    const workspaceKeys = Object.keys(workspace);
+    for (const key of workspaceKeys) {
+        // If the target property is an array, return its key and its contents.
+        if (Array.isArray(workspace[key])) {
+            return {
+                key: curKey,
+                value: workspace[key]
+            };
+        } else if (workspace[key] instanceof Object) {
+            // If the target property is an object, go one level in.
+            if (workspace.hasOwnProperty(key)) {
+                const newValue = getPropertyFromWorkspace(targetProp, workspace[key], key);
+                if (newValue !== null) {
+                    return newValue;
+                }
+            }
+        }
+    }
+
+    return null;
+}
+
+function includeDependencies(pkgJson: any, context: SchematicContext, tree: Tree, dependenciesKey: string) {
+    Object.keys(pkgJson.dependencies).forEach(pkg => {
+        const version = pkgJson.dependencies[pkg];
+        switch (pkg) {
+            case 'hammerjs':
+                logIncludingDependency(context, pkg, version);
+                addPackageToPkgJson(tree, pkg, version, dependenciesKey);
+
+                const workspace = getWorkspace(tree);
+                const targetProp = getPropertyFromWorkspace('scripts', workspace);
+                const mainTsPath = 'src/main.ts';
+                const hammerImport = 'import \'hammerjs\';\n';
+                const mainTsContent = tree.read(mainTsPath).toString();
+                // if there are no elements in the architect.build.options.scripts array that contain hammerjs
+                // and main.ts does not contain an import with hammerjs
+                if (!targetProp.value.some(el => el.includes('hammerjs')) && !mainTsContent.includes(hammerImport)) {
+                    // import hammerjs in the main.ts file
+                    const contents = hammerImport + mainTsContent;
+                    tree.overwrite(mainTsPath, contents);
+                }
+
+                break;
+            default:
+                logIncludingDependency(context, pkg, version);
+                addPackageToPkgJson(tree, pkg, version, dependenciesKey);
+                break;
+        }
+    });
+}
+
+/**
+ * Add an item to an angular.json section, within the architect
  * @param workspace Angular Workspace Schema (angular.json)
  * @param key Architect tool key to add option to
  */
-function addHammerToAngularWorkspace(workspace: WorkspaceSchema, key: string): boolean {
+function addItemToAngularWorkspace(workspace: WorkspaceSchema, key: string, item: string): boolean {
     const currentProjectName = workspace.defaultProject;
     if (currentProjectName) {
         if (!workspace.projects[currentProjectName].architect) {
@@ -83,8 +124,8 @@ function addHammerToAngularWorkspace(workspace: WorkspaceSchema, key: string): b
         if (!workspace.projects[currentProjectName].architect[key].options.scripts) {
             workspace.projects[currentProjectName].architect[key].options.scripts = [];
         }
-        if (!workspace.projects[currentProjectName].architect[key].options.scripts.includes(hammerJsMinAddress)) {
-            workspace.projects[currentProjectName].architect[key].options.scripts.push(hammerJsMinAddress);
+        if (!workspace.projects[currentProjectName].architect[key].options.scripts.includes(item)) {
+            workspace.projects[currentProjectName].architect[key].options.scripts.push(item);
             return true;
         }
